requirements.txt
dearpygui==2.0.0
pyserial==3.5
